refactor: split watch command into modular components
- Extract display logic to watch-display.internal.ts (328 lines)
- Extract input/keyboard handling to watch-input.internal.ts (110 lines)
- Extract adaptive scheduler to watch-scheduler.internal.ts (40 lines)
- Reduce watch.ts from 512 to 82 lines for better maintainability
- Each module now has single responsibility
